move reflection logic to utils package
