+++ .trunk/trunk.yaml
@@ -31,7 +31,7 @@ lint:
     - markdownlint@0.43.0
     - osv-scanner@1.9.2
     - prettier@3.4.2
-    - renovate@39.88.0
+    - renovate@39.90.2
     - trufflehog@3.88.0
     - yamllint@1.35.1
